Updated Submission model
Changed from TextField to CharField. Name is a tag. I think 80
characters is more than enough for it. Commit is a git commit, which is
sha1 sum 40 chars long.

Added indexes to names to make search faster. This is important at least
for Submission model, where we'll have a lot of records.

Change-Id: I3468ba732d36daeab25b709101d727e17706701d
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>