Change order of static files dires.
If there are files with the same name in STATICFILES_DIRS, django
manage.py collectstatic will only pick up the first one. So change
the order to put iris's static file dir at the top.

Lib which required by others should be put in front. For example
jquery is used by bootstrap and datatable, so it need be put
before them.

Fixes: 1860
Change-Id: I37c1ff9446098dddf943df270ce899e4496167b0
Signed-off-by: Huang Hao <hao.h.huang@intel.com>