Use CDN version of static files.
The site cdnjs.com maintained thousands of front-end libs. It's
a very popular site in recent years. So we change to use the
versions it maintained and void to package those static files
by ourselves.

jquery-multiselect was in cdnjs also, but it disappeared recently.
So we don't change multiselect in this patch.

Fixes: 1912
Change-Id: Ibf5cd48c01a491eed17821de59f8409baa36329c
Signed-off-by: Huang Hao <hao.h.huang@intel.com>