Remove 0001 migration.
We didn't use migrate in the frist place. And we already use
syncdb to create all tables for core application. Then 0001 is
redundant, it tries to create all tables we already have, which
will raise error.

Simplely skip the frist migrate and start from 0002. This solution
had been tested on iris-pptest server.

Change-Id: I29e524f81163e7d05f0380f2914ca98f748029ae