Use requests instead of urllib2
Requests allow you to send HTTP/1.1 requests. You can add headers, form data,
multipart files, and parameters with simple Python dictionaries, and access the
response data in the same way. It's powered by httplib and urllib3, it is much
more simpler and cleaner.
Urllib2 requires an enormous amount of work to perform the simplest of tasks.

Change-Id: Ibdacaebb9c450a8343a6ad5f48c2b210469bce5c
Signed-off-by: gaoxuesx <xuesongx.gao@intel.com>