Revert deleting migrations/0001_initial.py
When we install the latest version of iris-core rather than
upgrading from old version, it will complain that some database
table doesn't exist.

Because when we upgrade, the old tables are already there, all
migrations after 0001 can be run successfully. If we install
from scrach, iris.core app is managered by south, syncdb command
won't create any tables but use this 0001 migration to create
initial tables. If this file isn't there, all migration after
0001 will fail.

So we have to add it back, and keep it in source control in all
future versions.

Change-Id: Ia8719e46f4228a59350e1e72e3f111da54eac3f4