Fixes bug of scm_update API
In scm_update API the method read() of file obj is called twice,
at the twice time, it returns empty string, and this will delete
all data in IRIS db.
The paramater "content" of parse_block function must not be empty.
Add test cases for scm update and check API.

Change-Id: Ib7716e65ec646119b00735b75dfc581a5aed1018
Signed-off-by: gaoxuesx <xuesongx.gao@intel.com>