Add explicit csrf token for cache view
Since gittrees and packages including sign in post form, for security,
we need to add csrf token in it, so the two pages should render with
csrf token. However, we used cache for them, in django doc, it mentions:
However, if you use cache decorators on individual views, the CSRF
middleware will not yet have been able to set the Vary header or the
CSRF cookie, and the response will be cached without either one. In this
case, on any views that will require a CSRF token to be inserted you
should use the django.views.decorators.csrf.csrf_protect() decorator
first. see detais at:
https://docs.djangoproject.com/en/1.7/ref/contrib/csrf/#caching

Change-Id: I528e947164a6bff05ccb866b8820b5f3e12fa390
Fixes: #2220